Replace tomllib with toml
The tomllib module is only part of the std lib of Pythoon since 3.11.

This is way too bleeding edge so we will have to rely on the pypi package toml.

I will look into getting around this by just using Docker to manage Python interpreter versions.

That seems heavy handed to me.