Add paths module.
There is a lot of directory and file path related functionality in this project.

I was starting to see repetitive path string building across modules so I have
split this out into its own module.

There is likely other areas like URLs where this can be done.

I am aware that there is other paths I have 'missed'. I want to add this module
and a small set of changes and incrementally build on that as it is easier to review.

I will add the 'missing' paths to the paths.py module in due time.