Add functionality finding feature.
This feature allows you to specify regular expression string patterns in the config to search downloaded plugins for.

A report will then be generated with how many times each expression was found in each plugin.

There is also a sort key in the config that will sort in descending order plugins by the sort key.

For example if you had.

('wordpress-seo', {'$wpdb->prepare': 17, 'nopriv': 0})
('jetpack', {'$wpdb->prepare': 22, 'nopriv': 5})

You can use the sort key of 'nopriv' to get plugins that have the most unauthenticated endpoints to bubble to the top.

------------------------------------

Other changes.

- Config active installs defaults to 4M now so someone who runs this without reading the docs doesnt hammer the wordpress API/CDN.
- Added sort key to the config.
- Added very basic report generator to demo the functionality finder feature.

Nearing good enough for a v1.0.0 release.