Add report module.
The CLI should not be responsible for writing the report.

Write now the report format is very simple. I can see it getting complicated in the future.

Potentially even becoming an SQLite DB or a DB of some form.

Therefore I am moving the report writing functionality out into its own module.

There isn't much else to say right now about it.

It writes a badly structured file with a header and contents that are hard to parse/split.

I will address this for the v1.0.0 release.