Do not extract plugin zip files that have already been extracted.
If you modify the config to a low active install like 200k. You will end up with a lot
of zip files on your computer.

If you are changing your config settings to add and remove strings/regular expression patterns to search for. It is
highly expensive to keep extracting every single plugin again every time you want to check for some strings.

.i.e. there is no point in doing run.py -e if you only want to use run.py -c UNLESS! a new plugin archive has been downloaded.

This change will only extract plugin zip files if the directory they'd create does not already exist.

So it is possible that you could modify the existing files and the CLI will not reextract the zip file.

I will add a flag to do this for the v1.0.0. release.