Remove find functionality feature.
The find feature was fun to write but it was not implemented well.

It was essentially a find command piped into grep with regular expressions but handling less
edge cases.

The main issue I had when I first wrote this code was that it was difficult to download and
extract a large amount of wordpress plugins matching a certain criteria to a single directory.

I have pushed the code that has the find and report generating feature to the 'find' branch.

There are much better SAST tools such as SonarQube and Semgrep.

https://semgrep.dev/
https://www.sonarsource.com/products/sonarqube/

I'd recommend using this Python CLI to collect the plugins you need and then do the actually code analysis with the above tools.